fix bug and submit final solution

+++ src/main/java/org/gonchar/leetcode/AverageSalaryExcludingMinMax.java
@@ -7,7 +7,7 @@ public class AverageSalaryExcludingMinMax {
         double min = 10000000.0; // greater than any salary according to constraint
         double max = 10.0; // less than any salary according to constraint
         for (int i=0; i<salary.length; i++) {
-            double current = salary[i] / (salary.length - 2);
+            double current = ((double) salary[i]) / (salary.length - 2);
             if (max < current) {
                 max = current;
             }
